setting up unit tests fixed info data return

+++ buzzex-dot-net/buzzez-dot-net/Contracts/TradingPairBase.cs
@@ -33,7 +33,7 @@ namespace buzzez_dot_net.Contracts
         public decimal MinTotal { get; set; }
 
         [JsonProperty(PropertyName = "hidden")]
-        public decimal Hidden { get; set; }
+        public int Hidden { get; set; }
 
         [JsonProperty(PropertyName = "fee")]
         public decimal Fee { get; set; }
